Move `enek/feature_template/input_type.hpp` to `enek/feature_template/parsing/input_type.hpp`.
