+++ include/enek/feature_template/parsing/parse.hpp
@@ -2,7 +2,7 @@
 #define ENEK_FEATURE_TEMPLATE_PARSING_PARSE_HPP_INCLUDE_GUARD
 
 #include <enek/feature_template/parsing/ast.hpp>
-#include <enek/feature_template/input_type.hpp>
+#include <enek/feature_template/parsing/input_type.hpp>
 #include <filesystem>
 #include <iosfwd>
 #include <string>
@@ -11,13 +11,13 @@
 namespace Enek::FeatureTemplate::Parsing{
 
 Enek::FeatureTemplate::Parsing::AST
-parse(Enek::FeatureTemplate::InputType input_type,
+parse(Enek::FeatureTemplate::Parsing::InputType input_type,
       std::filesystem::path const &path,
       std::string const &text,
       std::ostream &os);
 
 Enek::FeatureTemplate::Parsing::AST
-parse(Enek::FeatureTemplate::InputType input_type,
+parse(Enek::FeatureTemplate::Parsing::InputType input_type,
       std::string const &text,
       std::ostream &os);
