+++ test/unit/feature_template/parsing/common.cpp
@@ -1,6 +1,6 @@
 #include "common.hpp"
 #include <enek/feature_template/parsing/parse.hpp>
-#include <enek/feature_template/input_type.hpp>
+#include <enek/feature_template/parsing/input_type.hpp>
 #include <gtest/gtest.h>
 #include <sstream>
 #include <string>
@@ -14,7 +14,9 @@ void testParse(
   namespace Parsing = Enek::FeatureTemplate::Parsing;
   std::ostringstream error_oss;
   Parsing::AST ast = Parsing::parse(
-    Enek::FeatureTemplate::InputType::word, text_to_be_parsed, error_oss);
+    Enek::FeatureTemplate::Parsing::InputType::word,
+    text_to_be_parsed,
+    error_oss);
   EXPECT_EQ(ast.succeed(), error_oss.str().empty());
   std::ostringstream dump_oss;
   ast.dumpXML(dump_oss);
